requirements.txt
absl-py==1.4.0
attrs==23.1.0
Automat==22.10.0
brotlipy==0.7.0
cachetools==4.2.4
certifi==2023.5.7
cffi @ file:///croot/cffi_1670423208954/work
chardet==3.0.4
charset-normalizer==3.1.0
clearml==1.10.4
click==8.1.7
Comet==3.1.0
constantly==15.1.0
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
Flask==2.2.5
Flask-Cors==4.0.0
fonttools==4.38.0
furl==2.1.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.54.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.12.0
onnxsim==0.4.1
opencv-python==4.7.0.72
orderedmultidict==1.0.1
packaging==23.1
pandas==1.3.5
pathlib2==2.3.7.post1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.22.2
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
torch==1.13.1
torchaudio==0.12.0
torchvision==0.14.1
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.6.3
ultralytics==8.0.116
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==6.0
